requirements.txt
fastapi[standard]==0.123.0
freezegun==1.5.5
google-cloud-storage==3.8.0
google-genai==1.60.0
uvicorn==0.40.0
pydantic==2.12.5
python-dotenv==1.2.1
python-multipart==0.0.22
pytest==9.0.2
werkzeug==3.1.5
Pillow==11.0.0
